Content generator repaired
