+++ index.html
@@ -37,6 +37,6 @@
       src="https://cdn.jsdelivr.net/npm/p5@0.10.2/lib/p5.js"
     ></script>
     <script type="module" src="./javascript/p5setup.js"></script>
-    <script type="module" src="./generate.js"></script>
+    <script type="module" src="./javascript/generate.js"></script>
   </body>
 </html>
